feat: different equipment icons for utility types
