+++ src/app/data-management/account-facilities/facility-data/facility-energy-uses/facility-energy-use-group/facility-energy-use-group.component.html
@@ -73,7 +73,7 @@
                                         Name
                                     </th>
                                     <th class="w-25">
-                                        Energy Source
+                                        Industrial System 
                                     </th>
                                     <th class="w-25">
                                         Utility Types
