+++ src/app/data-management/account-facilities/facility-data/facility-energy-uses/results/facility-energy-uses-group-summary/facility-energy-uses-group-summary-table/facility-energy-uses-group-summary-table.component.html
@@ -28,7 +28,9 @@
         <tr>
             <td class="nowrap">
                 <a class="click-link" (click)="goToEquipment(groupData.equipmentGuid)">
-                    <span class="fa fa-plug"></span>
+                    @for(icon of groupData.icons; track icon) {
+                    <span class="fa-solid" [class]="icon + ' me-1'"></span>
+                    }
                     {{groupData.equipmentName}}
                 </a>
             </td>
