+++ src/app/data-management/account-facilities/facility-data/facility-energy-uses/results/facility-energy-uses-summary/facility-energy-uses-summary-table/facility-energy-uses-summary-table.component.html
@@ -26,6 +26,9 @@
         <tr>
             <td>
                 <a class="click-link" (click)="goToGroup(groupData.groupId)">
+                    @for(icon of groupData.groupIcons; track icon) {
+                    <span class="fa-solid" [class]="icon + ' me-1'"></span>
+                    }
                     {{groupData.groupName}}
                 </a>
             </td>
