+++ src/app/shared/sharedHelperFunctions.ts
@@ -2,6 +2,7 @@ import { ConvertValue } from "../calculations/conversions/convertValue";
 import { MeterPhase, MeterSource } from "../models/constantsAndTypes";
 import { DetailDegreeDay, WeatherDataSelection } from "../models/degreeDays";
 import { IdbFacility } from "../models/idbModels/facility";
+import { IdbFacilityEnergyUseEquipment } from "../models/idbModels/facilityEnergyUseEquipment";
 import { IdbUtilityMeter } from "../models/idbModels/utilityMeter";
 import { FuelTypeOption } from "./fuel-options/fuelTypeOption";
 import { StationaryOtherEnergyOptions } from "./fuel-options/stationaryOtherEnergyOptions";
@@ -239,4 +240,37 @@ export function getWeatherSearchFromFacility(facility: IdbFacility): string {
     } else if (facility.zip) {
         return facility.zip;
     }
+}
+
+
+export type EnergyUseIcons = 'fa-plug-circle-bolt' | 'fa-fire-flame-curved' | 'fa-gas-pump' | 'fa-box';
+export function getEnergyUseSourceIcons(equipment: IdbFacilityEnergyUseEquipment): Array<EnergyUseIcons> {
+    let icons: Array<EnergyUseIcons> = [];
+    if (!equipment || !equipment.utilityData || equipment.utilityData.length === 0) {
+        return ['fa-box'];
+    } else {
+        equipment.utilityData.forEach(utility => {
+            switch (utility.energySource) {
+                case "Electricity":
+                    if (!icons.includes('fa-plug-circle-bolt')) {
+                        icons.push('fa-plug-circle-bolt');
+                    }
+                    break;
+                case "Natural Gas":
+                    if (!icons.includes('fa-fire-flame-curved')) {
+                        icons.push('fa-fire-flame-curved');
+                    }
+                    break;
+                case "Other Fuels":
+                    if (!icons.includes('fa-gas-pump')) {
+                        icons.push('fa-gas-pump');
+                    }
+                    break;
+            }
+        });
+    }
+    if (icons.length === 0) {
+        icons.push('fa-box');
+    }
+    return icons;
 }
